# [LEP-1173] perf(eventstore/reboot): support event name column filter, optimize reboot events store reads