--- a/components/os/component.go
+++ b/components/os/component.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"os"
 	"runtime"
-	"sort"
 	"sync"
 	"time"
 
@@ -24,7 +23,6 @@ import (
 	"github.com/leptonai/gpud/pkg/eventstore"
 	"github.com/leptonai/gpud/pkg/file"
 	pkghost "github.com/leptonai/gpud/pkg/host"
-	hostevents "github.com/leptonai/gpud/pkg/host/events"
 	"github.com/leptonai/gpud/pkg/kmsg"
 	"github.com/leptonai/gpud/pkg/log"
 	"github.com/leptonai/gpud/pkg/process"
@@ -55,9 +53,8 @@ type component struct {
 	ctx    context.Context
 	cancel context.CancelFunc
 
-	rebootEventStore hostevents.RebootsStore
-	eventBucket      eventstore.Bucket
-	kmsgSyncer       *kmsg.Syncer
+	eventBucket eventstore.Bucket
+	kmsgSyncer  *kmsg.Syncer
 
 	countProcessesByStatusFunc           func(ctx context.Context) (map[string][]process.ProcessStatus, error)
 	zombieProcessCountThresholdDegraded  int
@@ -94,8 +91,6 @@ func New(gpudInstance *components.GPUdInstance) (components.Component, error) {
 	c := &component{
 		ctx:    cctx,
 		cancel: ccancel,
-
-		rebootEventStore: gpudInstance.RebootEventStore,
 
 		countProcessesByStatusFunc:           process.CountProcessesByStatus,
 		zombieProcessCountThresholdDegraded:  defaultZombieProcessCountThresholdDegraded,
@@ -174,55 +169,15 @@ func (c *component) LastHealthStates() apiv1.HealthStates {
 }
 
 func (c *component) Events(ctx context.Context, since time.Time) (apiv1.Events, error) {
-	if c.eventBucket == nil && c.rebootEventStore == nil {
+	if c.eventBucket == nil {
 		return nil, nil
 	}
 
-	var events apiv1.Events
-	if c.eventBucket != nil {
-		componentEvents, err := c.eventBucket.Get(ctx, since)
-		if err != nil {
-			return nil, err
-		}
-
-		if len(componentEvents) > 0 {
-			events = make(apiv1.Events, 0, len(componentEvents))
-			for _, ev := range componentEvents {
-				// to prevent duplicate events
-				// since "reboot" events and "os" events
-				// share the same event store bucket "os"
-				if ev.Name == hostevents.RebootEventName {
-					continue
-				}
-				events = append(events, ev.ToEvent())
-			}
-		}
+	evs, err := c.eventBucket.Get(ctx, since)
+	if err != nil {
+		return nil, err
 	}
-
-	// for now,
-	// reboot events are recorded in the "os" bucket
-	// until we migrate, this method manually selects
-	// only the "reboot" events from the "os" bucket
-	// thus there should be no overlap between "eventBucket" and "rebootEventStore"
-	if c.rebootEventStore != nil {
-		rebootEvents, err := c.rebootEventStore.Get(ctx, since)
-		if err != nil {
-			return nil, err
-		}
-		if len(rebootEvents) > 0 {
-			events = append(events, rebootEvents.Events()...)
-		}
-	}
-
-	if len(events) == 0 {
-		return nil, nil
-	}
-
-	sort.Slice(events, func(i, j int) bool {
-		return events[i].Time.Time.After(events[j].Time.Time)
-	})
-
-	return events, nil
+	return evs.Events(), nil
 }
 
 func (c *component) Close() error {