--- a/pkg/host/events/doc.go
+++ b/pkg/host/events/doc.go
@@ -1,0 +1,2 @@
+// Package events provides the event store for the host.
+package events