--- a/pkg/host/events/reboots.go
+++ b/pkg/host/events/reboots.go
@@ -3,7 +3,6 @@ package events
 import (
 	"context"
 	"fmt"
-	"sort"
 	"time"
 
 	apiv1 "github.com/leptonai/gpud/api/v1"
@@ -40,29 +39,27 @@ var _ RebootsStore = &rebootsStore{}
 type rebootsStore struct {
 	getTimeNowFunc    func() time.Time
 	getLastRebootTime func(context.Context) (time.Time, error)
-	eventStore        eventstore.Store
+	bucket            eventstore.Bucket
 }
 
-func NewRebootsStore(eventStore eventstore.Store) RebootsStore {
+func NewRebootsStore(bucket eventstore.Bucket) RebootsStore {
 	return &rebootsStore{
 		getTimeNowFunc:    func() time.Time { return time.Now().UTC() },
 		getLastRebootTime: host.LastReboot,
-		eventStore:        eventStore,
+		bucket:            bucket,
 	}
 }
 
 func (s *rebootsStore) Record(ctx context.Context) error {
-	return recordEvent(ctx, s.eventStore, s.getTimeNowFunc(), s.getLastRebootTime)
-}
-
-func recordEvent(ctx context.Context, eventStore eventstore.Store, now time.Time, getLastRebootTime func(context.Context) (time.Time, error)) error {
-	currentBootTime, err := getLastRebootTime(ctx)
+	currentBootTime, err := s.getLastRebootTime(ctx)
 	if err != nil {
 		return err
 	}
 
 	// if now - event time > retention (too old), then skip
-	if now.Sub(currentBootTime) >= eventstore.DefaultRetention {
+	now := s.getTimeNowFunc()
+	elapsedSinceLastReboot := now.Sub(currentBootTime)
+	if elapsedSinceLastReboot >= eventstore.DefaultRetention {
 		log.Logger.Debugw("skipping reboot event", "time_since", time.Since(currentBootTime), "retention", eventstore.DefaultRetention)
 		return nil
 	}
@@ -74,15 +71,9 @@ func recordEvent(ctx context.Context, eventStore eventstore.Store, now time.Time
 		Message: fmt.Sprintf("system reboot detected %v", currentBootTime),
 	}
 
-	bucket, err := eventStore.Bucket(RebootBucketName, eventstore.WithDisablePurge())
-	if err != nil {
-		return err
-	}
-	defer bucket.Close()
-
 	// to prevent duplicate events
 	// in case "uptime -s" returns outdated but already recorded timestamp
-	existing, err := bucket.Find(ctx, curRebootEvent)
+	existing, err := s.bucket.Find(ctx, curRebootEvent)
 	if err != nil {
 		return err
 	}
@@ -90,14 +81,14 @@ func recordEvent(ctx context.Context, eventStore eventstore.Store, now time.Time
 		return nil
 	}
 
-	prevRebootEvent, err := bucket.Latest(ctx)
+	prevRebootEvent, err := s.bucket.Latest(ctx)
 	if err != nil {
 		return err
 	}
 
 	// no previous reboot event
 	if prevRebootEvent == nil {
-		return bucket.Insert(ctx, curRebootEvent)
+		return s.bucket.Insert(ctx, curRebootEvent)
 	}
 
 	// previous reboot event happened after the current reboot event
@@ -107,47 +98,18 @@ func recordEvent(ctx context.Context, eventStore eventstore.Store, now time.Time
 	}
 
 	// reboot already recorded in the last minute, thus skip
-	elapsed := currentBootTime.Sub(prevRebootEvent.Time)
-	if elapsed > 0 && elapsed < time.Minute {
+	rebootDelta := currentBootTime.Sub(prevRebootEvent.Time)
+	if rebootDelta > 0 && rebootDelta < time.Minute {
 		return nil
 	}
 
 	// reboot not recorded in the last minute, thus record
-	return bucket.Insert(ctx, curRebootEvent)
+	return s.bucket.Insert(ctx, curRebootEvent)
 }
 
 func (s *rebootsStore) Get(ctx context.Context, since time.Time) (eventstore.Events, error) {
-	return getEvents(ctx, s.eventStore, since)
-}
-
-func getEvents(ctx context.Context, eventStore eventstore.Store, since time.Time) (eventstore.Events, error) {
-	rebootBucket, err := eventStore.Bucket(RebootBucketName, eventstore.WithDisablePurge())
-	if err != nil {
-		return nil, err
-	}
-	defer rebootBucket.Close()
-
 	// we used the same bucket "os" for both reboot and os events
 	// until we migrate, we need manual filtering
 	// otherwise, we will get non-reboot events from the "os" component kmsg watcher
-	allOSEvents, err := rebootBucket.Get(ctx, since)
-	if err != nil {
-		return nil, err
-	}
-
-	all := make(eventstore.Events, 0, len(allOSEvents)/2)
-	for _, ev := range allOSEvents {
-		// The returned events should NOT include other events from the "os" component (e.g., kmsg watcher).
-		if ev.Name != RebootEventName {
-			continue
-		}
-		all = append(all, ev)
-	}
-
-	// The returned events are in the descending order of timestamp (latest event first).
-	sort.Slice(all, func(i, j int) bool {
-		return all[i].Time.After(all[j].Time)
-	})
-
-	return all, nil
+	return s.bucket.Get(ctx, since, eventstore.WithEventNamesToSelect(RebootEventName))
 }