… components/accelerator/nvidia/gpu-counts/component_test.go
… components/accelerator/nvidia/nccl/component_test.go
… components/accelerator/nvidia/peermem/component_test.go
… components/accelerator/nvidia/remapped-rows/component_test.go
… components/accelerator/nvidia/sxid/component_test.go
… components/accelerator/nvidia/xid/component_test.go
… components/cpu/component_test.go
… components/disk/component_test.go
… components/fuse/component_test.go
… components/memory/component_test.go
… components/os/component_test.go
… components/pci/component_test.go
… pkg/eventstore/database_benchmark_test.go
… pkg/eventstore/database_test.go
… pkg/eventstore/types_test.go
… pkg/host/events/reboots_benchmark_test.go
… pkg/host/events/reboots_test.go