- Don't select all events + if-statement to filter/select reboots ("os" bucket"), instead, just`select WHERE...`
- Don't create the SQL table for every single reboot event insert/reads (very inefficient!)

Combining all these, reboot events store read performance 3x.

And we will need the same eventstore helper for the future Infiniband flap detection.

```
                                  │ baseline.txt │             new.txt             │
                                  │    sec/op    │    sec/op     vs base           │
EventStore_GetEvents_Filtering-10   3.752m ± ∞ ¹   2.787m ± ∞ ¹  ~ (p=1.000 n=1) ²

                                  │ baseline.txt │             new.txt             │
                                  │    sec/op    │    sec/op     vs base           │
EventStore_GetEvents_Filtering-10   3.542m ± ∞ ¹   2.762m ± ∞ ¹  ~ (p=1.000 n=1) ²

                                  │ baseline.txt  │             new.txt              │
                                  │     B/op      │     B/op       vs base           │
EventStore_GetEvents_Filtering-10   3.674Mi ± ∞ ¹   2.375Mi ± ∞ ¹  ~ (p=1.000 n=1) ²

                                  │ baseline.txt │             new.txt             │
                                  │  allocs/op   │  allocs/op    vs base           │
EventStore_GetEvents_Filtering-10   62.52k ± ∞ ¹   50.03k ± ∞ ¹  ~ (p=1.000 n=1) ²
```

```
                                        │ /tmp/baseline1.txt │          /tmp/new1.txt          │
                                        │       sec/op       │    sec/op     vs base           │
RebootsStore_Record_NoExistingEvents-10         32.05µ ± ∞ ¹   10.99µ ± ∞ ¹  ~ (p=1.000 n=1) ²

                                        │ /tmp/baseline1.txt │          /tmp/new1.txt           │
                                        │        B/op        │     B/op       vs base           │
RebootsStore_Record_NoExistingEvents-10        6.252Ki ± ∞ ¹   2.846Ki ± ∞ ¹  ~ (p=1.000 n=1) ²

                                        │ /tmp/baseline1.txt │         /tmp/new1.txt          │
                                        │     allocs/op      │  allocs/op   vs base           │
RebootsStore_Record_NoExistingEvents-10         160.00 ± ∞ ¹   73.00 ± ∞ ¹  ~ (p=1.000 n=1) ²
```

```
                              │ /tmp/baseline2.txt │          /tmp/new2.txt          │
                              │       sec/op       │    sec/op     vs base           │
RebootsStore_Get_SmallData-10        4980.5µ ± ∞ ¹   812.7µ ± ∞ ¹  ~ (p=1.000 n=1) ²

                              │ /tmp/baseline2.txt │          /tmp/new2.txt           │
                              │        B/op        │     B/op       vs base           │
RebootsStore_Get_SmallData-10       4060.2Ki ± ∞ ¹   483.3Ki ± ∞ ¹  ~ (p=1.000 n=1) ²

                              │ /tmp/baseline2.txt │          /tmp/new2.txt          │
                              │     allocs/op      │  allocs/op    vs base           │
RebootsStore_Get_SmallData-10         66.16k ± ∞ ¹   11.06k ± ∞ ¹  ~ (p=1.000 n=1) ²
```

```
                              │ /tmp/baseline3.txt │          /tmp/new3.txt          │
                              │       sec/op       │    sec/op     vs base           │
RebootsStore_Get_LargeData-10        25.103m ± ∞ ¹   4.088m ± ∞ ¹  ~ (p=1.000 n=1) ²

                              │ /tmp/baseline3.txt │          /tmp/new3.txt           │
                              │        B/op        │     B/op       vs base           │
RebootsStore_Get_LargeData-10       23.098Mi ± ∞ ¹   2.903Mi ± ∞ ¹  ~ (p=1.000 n=1) ²

                              │ /tmp/baseline3.txt │          /tmp/new3.txt          │
                              │     allocs/op      │  allocs/op    vs base           │
RebootsStore_Get_LargeData-10        330.17k ± ∞ ¹   55.07k ± ∞ ¹  ~ (p=1.000 n=1) ²
```

```
                                      │ /tmp/baseline4.txt │          /tmp/new4.txt          │
                                      │       sec/op       │    sec/op     vs base           │
RebootsStore_Record_ExistingEvents-10         30.79µ ± ∞ ¹   10.90µ ± ∞ ¹  ~ (p=1.000 n=1) ²

                                      │ /tmp/baseline4.txt │          /tmp/new4.txt           │
                                      │        B/op        │     B/op       vs base           │
RebootsStore_Record_ExistingEvents-10        6.253Ki ± ∞ ¹   2.846Ki ± ∞ ¹  ~ (p=1.000 n=1) ²

                                      │ /tmp/baseline4.txt │         /tmp/new4.txt          │
                                      │     allocs/op      │  allocs/op   vs base           │
RebootsStore_Record_ExistingEvents-10         160.00 ± ∞ ¹   73.00 ± ∞ ¹  ~ (p=1.000 n=1) ²
```

c.f., https://github.com/leptonai/gpud/pull/967