@eahydra 2025-07-07
This is a case with more reads than writes. From your benchmark results, there is actually no essential change from baseline to new. They are all very slow because they are still SQL operations. Should we build a cache directly based on the latest DefaultRetentionPeriod event? Add a goroutine to refresh the cache periodically in the background. Is this more efficient?