@gyuho 2025-07-07
Since reboot bucket/table will remain pretty small, this is not the critical bottleneck, this is to mainly remove unnecessary filtering. (you don't need if-statement, when you can just use `WHERE` statement).